Split the VSTS DI boostrapper from the Actors entry point into it's own module
