+++ src/Sierra.Actor/Program.cs
@@ -5,15 +5,9 @@
     using System.Threading.Tasks;
     using Autofac;
     using Autofac.Integration.ServiceFabric;
-    using Common;
+    using Common.DependencyInjection;
     using Eshopworld.Telemetry;
-    using Microsoft.Azure.KeyVault;
-    using Microsoft.Azure.Services.AppAuthentication;
-    using Microsoft.Extensions.Configuration;
-    using Microsoft.Extensions.Configuration.AzureKeyVault;
     using Microsoft.TeamFoundation.SourceControl.WebApi;
-    using Microsoft.VisualStudio.Services.Common;
-    using Microsoft.VisualStudio.Services.WebApi;
 
     internal static class Program
     {
@@ -24,33 +18,10 @@
         {
             try
             {
-                var configBuilder = new ConfigurationBuilder().AddAzureKeyVault(
-                    @"https://esw-tooling-ci.vault.azure.net/",
-                    new KeyVaultClient(new KeyVaultClient.AuthenticationCallback(new AzureServiceTokenProvider().KeyVaultTokenCallback)),
-                    new DefaultKeyVaultSecretManager());
-
                 var builder = new ContainerBuilder();
+                builder.RegisterModule(new VstsModule {Vault = @"https://esw-tooling-ci.vault.azure.net/"});
+
                 builder.RegisterServiceFabricSupport();
-                var config = configBuilder.Build();
-
-                builder.RegisterInstance(config)
-                       .As<IConfigurationRoot>()
-                       .SingleInstance();
-
-                var vstsConfig = new VstsConfiguration();
-                config.Bind(vstsConfig);
-
-                builder.RegisterInstance(vstsConfig)
-                       .SingleInstance();
-
-                builder.Register(c => new VssBasicCredential(string.Empty, c.Resolve<VstsConfiguration>().VstsPat))
-                       .SingleInstance();
-
-                builder.Register(c => new VssConnection(new Uri(c.Resolve<VstsConfiguration>().VstsBaseUrl), c.Resolve<VssBasicCredential>()))
-                       .InstancePerDependency();
-
-                builder.Register(c => c.Resolve<VssConnection>().GetClient<GitHttpClient>())
-                       .InstancePerDependency();
 
                 builder.RegisterActor<TenantActor>();
                 builder.RegisterActor<LockerActor>();
