+++ src/Sierra.Actor/Sierra.Actor.csproj
@@ -16,6 +16,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Autofac" Version="4.8.1" />
     <PackageReference Include="Autofac.ServiceFabric" Version="2.0.0" />
     <PackageReference Include="Eshopworld.Telemetry" Version="1.0.0" />
     <PackageReference Include="Microsoft.Azure.KeyVault" Version="2.3.2" />
