+++ src/Tests/Sierra.Api.Tests/Sierra.Api.Tests.csproj
@@ -21,7 +21,7 @@
     <PackageReference Include="Eshopworld.DevOps" Version="1.0.5" />
     <PackageReference Include="Eshopworld.Tests.Core" Version="1.0.0" />
     <PackageReference Include="IdentityModel" Version="3.7.0-preview1" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.0.2" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.1.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.7.2" />
     <PackageReference Include="xunit" Version="2.3.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
